+++ assets.lua
@@ -107,6 +107,25 @@ Assets = GLOBAL.JoinArrays(
 		"pineapple_seeds",
 
 		"thunder_pinecone",
+
+		--Tea images.
+	    "whitetea",
+	    "petaltea",
+	    "mixedpetaltea",
+	    "evilpetaltea",
+	    "floraltea",
+	    "skypetaltea",
+	    "daturatea",
+	    "cloudfruittea",
+	    "greenertea",
+	    "berrytea",
+	    "berrypulptea",
+	    "cottontea",
+	    "goldtea",
+	    "candytea",
+
+	    --"chaitea",
+	    --"oolongtea",
 	},
 
 	-- These are .tex's without a matching atlas.