+++ code/prefabs/tea.lua
@@ -12,7 +12,7 @@ local cfg = Configurable("HOTBEVERAGE", "TEA")
 -- Basic tea data, as used by MakeBeverage.
 --]]
 local basic_tea_data = {
-	bank = "food",
+	bank = "tea",
 
 	perish_time = cfg:GetConfig("PERISH_TIME"),
 	heat_capacity = cfg:GetConfig("HEAT_CAPACITY"),
@@ -26,11 +26,11 @@ local tea_data = {
 		name = "greentea",
 
 		assets = {
-			Asset("ANIM", "anim/cook_pot_food.zip"),
+			Asset("ANIM", "anim/tea.zip"),
 		},
 
-		build = "cook_pot_food",
-		anim = "fruitmedley",
+		build = "tea",
+		anim = "greentea",
 		inventory_atlas = "images/inventoryimages/greentea.xml",
 
 		health = TUNING.HEALING_SMALL,
@@ -41,11 +41,11 @@ local tea_data = {
 
 	blacktea = MergeMaps(basic_tea_data, {
 		assets = {
-			Asset("ANIM", "anim/cook_pot_food.zip"),
+			Asset("ANIM", "anim/tea.zip"),
 		},
 
-		build = "cook_pot_food",
-		anim = "fruitmedley",
+		build = "tea",
+		anim = "tea",
 		inventory_atlas = "images/inventoryimages/blacktea.xml",
 
 		health = TUNING.HEALING_TINY,
@@ -56,12 +56,21 @@ local tea_data = {
 
 	whitetea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
 		build = "cook_pot_food",
 		anim = "fruitmedley",
 		--inventory_atlas = "images/inventoryimages/whitetea.xml",
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/whitetea.xml",
+>>>>>>> stable
 
 		health = TUNING.HEALING_SMALL,
 		sanity = TUNING.SANITY_SMALL,
@@ -71,6 +80,7 @@ local tea_data = {
 
 	petaltea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -80,12 +90,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/petaltea.xml",
+
+		health = 5,
+		sanity = 5,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
 	evilpetaltea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -95,12 +117,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/evilpetaltea.xml",
+
+		health = 8,
+		sanity = -10,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
 	mixedpetaltea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -110,12 +144,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/mixedpetaltea.xml",
+
+		health = 5,
+		sanity = -2,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
 	floraltea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -125,12 +171,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "floraltea",
+		inventory_atlas = "images/inventoryimages/floraltea.xml",
+
+		health = 0,
+		sanity = 10,
+>>>>>>> stable
 
 		temperature = 70,
 	}),
 
 	berrytea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -140,12 +198,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/berrytea.xml",
+
+		health = 5,
+		sanity = 10,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
 	berrypulptea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -155,12 +225,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/berrypulptea.xml",
+
+		health = 2,
+		hunger = 10,
+>>>>>>> stable
 
 		temperature = 40,
 	}),
 
 	skypetaltea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -170,12 +252,24 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/skypetaltea.xml",
+
+		health = -12,
+		sanity = 18,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
 	daturatea = MergeMaps(basic_tea_data, {
 		assets = {
+<<<<<<< HEAD
 			Asset("ANIM", "anim/cook_pot_food.zip"),
 		},
 
@@ -185,10 +279,22 @@ local tea_data = {
 
 		health = 0,
 		sanity = 0,
+=======
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/daturatea.xml",
+
+		health = 20,
+		sanity = -10,
+>>>>>>> stable
 
 		temperature = 60,
 	}),
 
+<<<<<<< HEAD
 	greengreentea = MergeMaps(basic_tea_data, {
 		assets = {
 			Asset("ANIM", "anim/cook_pot_food.zip"),
@@ -203,6 +309,316 @@ local tea_data = {
 
 		temperature = 60,
 	}),
+=======
+	greenertea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/greenertea.xml",
+
+		health = 30,
+		sanity = -20,
+
+		temperature = 60,
+	}),
+
+	cloudfruittea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/cloudfruittea.xml",
+
+		hunger = 10,
+		health = 10,
+
+		temperature = 30,
+	}),
+
+	cottontea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/cottontea.xml",
+
+		hunger = 5,
+		sanity = 5,
+
+		temperature = 60,
+	}),
+
+	candytea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/candytea.xml",
+
+		health = -10,
+		sanity = 20,
+
+		temperature = 50,
+	}),
+
+	goldtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		inventory_atlas = "images/inventoryimages/goldtea.xml",
+
+		health = 0,
+		sanity = 20,
+
+		temperature = 70,
+	}),
+
+	chaitea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/chaitea.xml",
+
+		health = TUNING.HEALING_TINY,
+		sanity = TUNING.SANITY_SMALL,
+
+		temperature = 100,
+	}),
+
+	oolongtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/oolongtea.xml",
+
+		health = TUNING.HEALING_TINY,
+		sanity = TUNING.SANITY_SMALL,
+
+		temperature = 100,
+	}),
+
+	redmushroomtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/redmushroomtea.xml",
+
+		health = TUNING.HEALING_TINY,
+		sanity = -TUNING.SANITY_SMALL,
+		hunger = -3,
+
+		temperature = 100,
+	}),
+
+	bluemushroomtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/bluemushroomtea.xml",
+
+		health = -TUNING.HEALING_SMALL,
+		sanity = TUNING.SANITY_SMALL,
+		hunger = -2,
+
+		temperature = 100,
+	}),
+
+	greenmushroomtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/greenmushroomtea.xml",
+
+		health = -TUNING.HEALING_TINY,
+		sanity = TUNING.SANITY_MED,
+		hunger = -1,
+
+		temperature = 100,
+	}),		
+
+	redjellytea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/redjellytea.xml",
+
+		health = -3,
+		sanity = 30,
+		hunger = -30,
+
+		temperature = 100,
+	}),
+
+	bluejellytea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/bluejellytea.xml",
+
+		health = 30,
+		sanity = -30,
+		hunger = -3,
+
+		temperature = 100,
+	}),
+
+	greenjellytea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/greenjellytea.xml",
+
+		health = -30,
+		sanity = -3,
+		hunger = 30,
+
+		temperature = 100,
+	}),	
+
+	jellytea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/jellytea.xml",
+
+		health = 3,
+		sanity = 3,
+		hunger = 3,
+
+		temperature = 99,
+	}),	
+
+	redtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/redtea.xml",
+
+		--health = TUNING.HEALING_TINY,
+		--sanity = TUNING.SANITY_SMALL,
+
+		temperature = 100,
+	}),	
+
+	herbaltea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/herbaltea.xml",
+
+		health = 20,
+		sanity = 10,
+
+		temperature = 90,
+	}),	
+
+	marshmallowtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/marshmallowtea.xml",
+
+		health = -6,
+		sanity = 10,
+
+		temperature = 80,
+	}),
+
+	dragontea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/dragontea.xml",
+
+		health = -40,
+		sanity = -5,
+
+		temperature = 200,
+	}),	
+
+	ambrosiatea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/ambrosiatea.xml",
+
+		health = math.random(-40,20),
+		sanity = math.random(-40,20),
+		hunger = math.random(-40,20),
+
+		temperature = 100,
+	}),	
+
+	spoiledtea = MergeMaps(basic_tea_data, {
+		assets = {
+			Asset("ANIM", "anim/tea.zip"),
+		},
+
+		build = "tea",
+		anim = "tea",
+		--inventory_atlas = "images/inventoryimages/spoiledtea.xml",
+
+		health = -20,
+		sanity = -10,
+
+		temperature = 90,
+	}),
+
+>>>>>>> stable
 }