+++ code/resources/refining_recipebook.lua
@@ -23,4 +23,5 @@ return RecipeBook {
 	refined_petals = Recipe("petals", Ingredient("golden_petals", 1), 0),
 	refined_gold   = Recipe("goldnugget", Ingredient("golden_petals", 4), 1),
 	refined_ash    = Recipe("ash", Ingredient("cloud_jelly", 1), 0),
+	refined_logs   = Recipe("log", Ingredient("thunder_log", 1), 0),
 }